refactor and added gym
Added gym to the worker. It uses the official image and extends it by
adding its own functionality on top.
BUG: gym can not be imported directly, this is a bug in the base image
(which I have no control over). For now I fixed it by defining the
appropriate environment variable

moved the test worker to the actual worker, since it has matured enough.