automated build and setup of swarm
created automated build script for boot2docker.iso
Build script adds REX-Ray to boot2docker and adds a (configurable)
config file that drives it.
Further it generates a signed certificates with IP-SAN, to be consumed
by the registry server to allow image push / pull. The certificate is
also integrated into the image, so that all nodes in the swarm accept
the connection.

Further added a script that allows automated deployment of a base stack:
deploy a registry including persistent storage of images (using rexray)
and secrets for the certificate.
deploy portainer on all manager nodes, to get a nice graphical overview
over the swarm's current activity.