fixed all hickups + bughunt
I#ve spend the entire day trying to fix a bug that is none.
Essentially the reducer has multiple final episodes while collecting.
This is because each step can occur multiple times, which isnt a
problem per se.
However, for some reason the agent-environment pair is NOT
reproducible, thus producing multiple different results at every turn.
I have to get to the bottom of this, but suspect openAI gym to not be
consistent.